requirements.txt

psutil
ninja
accelerate
beartype
clean-fid
colorama
colorspacious
dacite
diffusers
e3nn
easydict
einops
h5py
hydra-core
imageio
imageio-ffmpeg
ipykernel
ipython
jaxtyping
lightning
lpips
matplotlib
matplotlib-inline
moviepy
natsort
openai
pandas
pillow
plyfile
rotary-embedding-torch
ruff
scikit-image
seaborn
tabulate
timm==1.0.15
tqdm
transformers
wandb
